backend/requirements.txt
# 后端依赖列表

# 核心框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic[email]>=2.5.0

# 数据库
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9  # PostgreSQL 驱动

# 认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.0

# AI服务
anthropic>=0.39.0
agno>=2.2.10
openai>=1.102.0

# HTML转Markdown
markdownify>=0.11.6

# 开发工具
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
black>=23.12.0
ruff>=0.1.0
mypy>=1.7.0
